rename: 일상/대피소 데이터 fetch 함수 동일하도록 디테일 수정
